Add ZSH plugins as submodules
It is technically possible to install these with apt-get but the fast
syntax highlighting requires a custom source added. It's better & easier
to just use these as git submodules and add the appropriate
configuration to the zshrc to make them work.

Auto-suggestions uses history and completion to make suggestions as you
type. Mostly it saves a bit of `^r` but that's still useful.

The syntax highlighting flags commands, arguments, etc. in different
colors. You don't really need that to survive but it is helpful to have
commands that aren't installed show up in `red` so that you know
something is missing. Same with tracking down unclosed strings and
stuff.

Testing this out as submodules.