Setup git config
Mostly this is a set of useful aliases as the spin git config is
already well configured for github. I'm just making things like
user name and email explicit because it makes me feel better.

Adding a placeholder gitignore - in the future I'll probably want
to be able to drop things like ctags files