Add some useful functions: zoom and visual block move
This adds two commands:

`:ZoomToggle`. This takes a split and zooms it to full screen, and then
when called a second time it restores it to the original size. It's
useful if you have some code + tests on screen and you want to focus on
just the one thing, to view long lines, or whatever

`:VisualUp` and `:VisualDown` can be called when there is a visual block
selection. They move the whole block up or down several lines. You can
write a slightly cheaper version but it breaks at the top and bottom of
the document. This approach is clever for those edgecases.