Rubocop Improvements. Length count of blocks and bugfix
This fixes an incorrect setting in the previous commit. The rule for
commas in multi-line arrays should have been `EnforcedStyleForMultiLine`
not `EnforcedStyle`.

This also adds a change in the way block parameters are computed. Multi
line arrays, hashes, and heredocs only increment the count by 1 rather
than 1 per line.

This also alphabetizes the configuration to make things easier to find
as this file grows.